🐛 fix(protocol): 通过能力配置处理 top_k
- capabilities: 新增 sampling.topK 并为 DeepSeek 预设启用
- adapters: OpenAI Chat 按能力透传，不支持时显式拒绝
- responses: 拒绝不支持 top_k 的标准 Responses 上游
- ir: 将 TopK 收紧为协议要求的整数类型
- test: 覆盖能力配置、整数校验和支持边界

+++ internal/ir/ir.go
@@ -74,10 +74,11 @@ type Tool struct {
 }
 
 type Params struct {
-	MaxTokens        *int            `json:"maxTokens,omitempty"`
-	Temperature      *float64        `json:"temperature,omitempty"`
-	TopP             *float64        `json:"topP,omitempty"`
-	TopK             *float64        `json:"topK,omitempty"`
+	MaxTokens   *int     `json:"maxTokens,omitempty"`
+	Temperature *float64 `json:"temperature,omitempty"`
+	TopP        *float64 `json:"topP,omitempty"`
+	// TopK is an integer in the Anthropic and Gemini APIs.
+	TopK             *int            `json:"topK,omitempty"`
 	Stop             []string        `json:"stop,omitempty"`
 	ReasoningEffort  *string         `json:"reasoningEffort,omitempty"`
 	Thinking         json.RawMessage `json:"thinking,omitempty"`
